Send lighting information for 1.14 and 1.15
